+++ src/enums/index.ts
@@ -1,2 +1,2 @@
 export * from "./level.enum";
-export * from "./console.enum";
+export * from "./type.enum";