+++ src/models/index.ts
@@ -1,2 +1,4 @@
-export * from "./logora.config";
 export * from "./logora.interface";
+export * from "./logora-output.interface";
+export * from "./logora-output-options.interface";
+