+++ src/module.ts
@@ -1,0 +1,4 @@
+
+
+export type { LogEntry } from "./core/log-entry";
+export type { ILogoraWriter } from "./core/logora-writer.interface";